Allowed Alternate Probe Desc & no space after \d oz
Software came in from Mexico, with a different wires file format that normal.

While it is possible that the wires file is completely manually edited,
it is also possible that it has been created using a newer version of software,
or software using different settings.

The script was updated to:
                               V
    place a space after eg 5 oz<node name>
    so it became:          5 0z <node name>

    Also replaces 100MilProbe with 100 mil etc