fix known issues with create_exe.bat.
Having to create a 32 bit and a 64 bit exe has been problematic.

The 64 bit was created for performance reasons etc, and the 32 bit
for compatability.

But due to me not adding the 32 bit version to the path, there were
some problems everytime I tried to account for this in the batch file.

Now the batch file:
    keeps a backup of the build files
    copies the build files into the folder at the start of the build
    resets the path after every tun.

    also, due to the common build files, it tends to run faster.